Always update storage size attr to ensure value in bytes
Previously devices which returned size attributes via udev
were left untouched, however the unit of the size attribute
is not always in bytes.  The _get_disk_size method looks up
the parent device and finds the sysfs queue/logical_block_size
parameter and uses that to update the block value into a
size in bytes.  Always use this method so the storage size
attribute returns device size in bytes.

Signed-off-by: Ryan Harper <ryan.harper@canonical.com>