testing: enable tox, traivs, flake8, and pytest
This adds a `make test` target that will execute tox. Tox in turn will
run py.test and flake8 against the repo.

Two fixes to the actual code to clean up flake8 issues, one change
to the network.py import around c-bindings so we do not have to build
the package for testing, and a few changes to test code to ensure
pylint passes. I choose this to speed up test/dev, rather than needing
deps and building.

Fixes #63