remove routes for an interface from libnl's cache when it does down
I was testing the live server installer for 18.04.3, just fooling
around, when I noticed that if I disabled and reenabled an interface,
subiquity didn't notice; the button on the network screen stubbornly
stayed at "[ Continue without networking ]". After staring at debug
output and the libnl3 source for a good long while, I figured out what
was going on, and it was very related to the bug I fixed in subiquity
recently where subiquity didn't notice the default route going away.

That bug was caused by the fact that when an interface goes down, there
are not netlink notifications for each route on that interface going
away.

For libnl3's route cache, this means that the default route hangs around
in the cache. When the interface comes back up and the route is added
again, libnl3 doesn't consider it to be new and doesn't notify probert
about it.

Fix this by manually removing all routes for an interface from libnl's
cached when we see it go down.

(Thank you for coming to my ted talk)