add wlan data to wlan devices as they are created
Previously this data was added when the first NL80211 message was seen
for a nic but it seems no such events are sent for mac80211_hwsim NICs
which otherwise would be useful for testing.

This change also conceivably fixes some bugs where a UI processes a wlan
NIC before the first NL80211 message is seen for it, but I've never had
any reports of that.