storage: skip partitions on device we are skipping
Subiquity installs are now failing when it tries to hand the probe
results to curtin and they contain partitions that are on a loopback
device.  Curtin tries to associate this partition to a disk, and that
disk doesn't exist because it has been filtered from the results.

So also filter those partitions.