Fix crash when PARTNAME is invalid utf-8
In GPT partitions, the PartitionName field is stored as a utf-16 string.
For some reason though, in udev, the value is not converted from utf-16
to utf-8. Instead, the MSB is dropped and the LSB is kept. e.g.,

 * letter A in UTF-16 is 0x0041. It becomes 0x41 (which is letter A)
 * e-ecute (é) in UTF-16 is 0x00E9. It becomes 0xE9
 * euro sign (€) in UTF-16 is 0x20AC. It becomes 0xAC
 * pound sign (£) in UTF-16 is 0x00A3. It becomes 0xA3

In the above examples, only the letter A results in a valid UTF-8
character.

When using pyudev, the values are all expected to be decodable using the
system's default encoding.

In probert, it can result in a crash when iterating over all the
properties of a block device, that are mapped by pyudev.

To work around the issue, we now stop iterating over all the values. We
iterate over the keys and only access the value if we need it.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>