record device maj:min in mount data
I want to make subiquity a bit more subtle in how it handles the install
media, which means it needs to understand the pre-existing mounts in the
live session a little better. Casper creates some mounts that are pretty
hard to decode from the information we save already, but if we record
the device major/minor we can pretty reliably map that to a partition.