utils: add function to read size of block device using DEVPATH
The existing read_sys_block_size_bytes functions forces us to guess
where to find a given device in the /sys/class/block hierarchy. While
most of the time, /sys/class/block/<device>/size seems to be the right
place to look, there have been reports of the file missing.

We now provide another way to read the size using the DEVPATH udev
variable instead. Hopefully, this should be a more accurate way to find
the information we are looking for.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>