utils: add optional logging if read_sys_devpath_size_bytes fails
If read_sys_devpath_size_bytes fails with ENOENT, we want to know
whether it's the directory that does not exist of the file named "size".

If the caller requests it, on error we now list the files contained in
the /sys/<devpath> directory.

If the directory does not exist, we will have a nested
FileNotFoundError exception.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>