Add way to probe firmware information
For NVMe/TCP, we will want Subiquity to determine if booting with the
boot FS on remote-storage is supported by the firmware. So far, it looks
like we might need to rely on the "bios-vendor" value so we now expose
this information from probert.

Since firmware related information should not change between two
different probert invocations now make it part of a "firmware" prober,
that Subiquity should realistically call only once.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>