nvme: run `nvme connect-all --nbft` before running the actual probes
When running in an environment where a NBFT is available, we want
probert to collect information about the remote NVMe drives too.

This can be done by calling the `nvme connect-all --nbft` command. That
said, running the command in the NVMe probe code itself will be too late
for other probes (e.g., blockdev, filesystem, ...) to pick-up
information about the remote NVMe drives.

Instead, we run the command before dispatching the probes ; in a
function called "activate_devices". In the long run, I would like to
move some code that is currently in the probes (e.g., vgchange
--activate=y) to the activate_devices function ; to avoid side effects
in the probe code.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>