_rtnetlinkmodule: filter out link objects having family=AF_INET6
Since libnl 3.6, a single link cache object can contain two rtnl_link object
for the same interface index:

 * one with family=AF_INET6
 * one with family=AF_UNSPEC or family=AF_BRIDGE

When this happens, calls to rtnl_link_get(cache, ifindex) suddently return one
or the other.

This leads to crashes when setting or unsetting link flags if we are not lucky
enough to get the right rtnl_link instance.

LP: #2105510

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>